feat: abort running c8 on Node.js < v10.0.0
